Fix CI again

+++ rcp/src/destination.rs
@@ -90,7 +90,7 @@ async fn handle_file_stream(
                 ));
             }
         }
-        throttle::get_file_iops_tokens(settings.chunk_size, file_header.size as u64).await;
+        throttle::get_file_iops_tokens(settings.chunk_size, file_header.size).await;
         let mut file = tokio::fs::File::create(&file_header.dst).await?;
         let copied = file_recv_stream.copy_to(&mut file).await?;
         if copied != file_header.size {
